#21 Refactor right click modals into separate components
